+++ app/Http/Controllers/Admin/TagsController.php
@@ -48,26 +48,10 @@ class TagsController extends Controller
     public function store(TagEditRequest $request)
     {
         $data = $request->validated();
-
-        $icon = $request->only('icon_id');
-        $data['icon_id'] = null;
-
-        if(isset($icon['icon_id'])) {
-
-            $icon = json_decode($icon['icon_id']);
-            if(!empty($icon) && !empty($icon[0]->id)) $data['icon_id'] = (int) $icon[0]->id;
-        }
-        
-        if(!empty($data['tags'])) $data['tags'] = PillFieldHelper::toArray($data['tags']);
-        if(!empty($data['asset_id'])) $data['asset_id'] = PillFieldHelper::toArray($data['asset_id'])[0];
-
-
-        
+        $data['user_id'] = Auth::id();
 
         $tag = Tag::create($data);
-        $tag->user()->attach(Auth::id());
-        if($data['tags']) $tag->tags()->attach($data['tags']);
-        
+  
         session()->flash('message', new MessageBag(['status' => 'success',
                                                     'message' => 'Good job! The tag was created.']));
   
@@ -86,18 +70,6 @@ class TagsController extends Controller
     {
         $tag = Tag::findOrFail($id);
 
-        $icon = $tag->icon()->first();
-        $tags = $tag->tags()->get();
-
-
-        if($icon) $tag->icon = json_encode([$icon->toArray()]);
-    
-        $tag->tags = PillFieldHelper::dbRowsToJson($tags->toArray(), 'id', 'name');
-
-        $asset = $tag->asset()->first();
-       
-        if($tag->asset_id) $tag->asset_id = json_encode([['id'=>$asset->id, 'value'=>$asset->title.' #'.$asset->asset_type]]);
-        
 
         return view('admin/taxonomies/tags/edit', ['tag' => $tag]);
     }
@@ -116,24 +88,8 @@ class TagsController extends Controller
         
         $tag = Tag::findOrFail($id);
 
-        $icon = $request->only('icon_id');
-        $data['icon_id'] = null;
-
-        if(isset($icon['icon_id'])) {
-
-            $icon = json_decode($icon['icon_id']);
-            if(!empty($icon) && !empty($icon[0]->id)) $data['icon_id'] = (int) $icon[0]->id;
-         }
-
-         if(!empty($data['tags'])) $data['tags'] = PillFieldHelper::toArray($data['tags']);
-         if(!empty($data['asset_id'])) $data['asset_id'] = PillFieldHelper::toArray($data['asset_id']);
-         
-         $data['asset_id'] = isset($data['asset_id'][0]) ? $data['asset_id'][0] : null;
-
-
         $tag->update($data);
-        $tag->tags()->sync($data['tags']);
- 
+
         session()->flash('message', new MessageBag(['status' => 'success',
                                                     'message' => 'Excellent! The tag was edited.']));
         
@@ -164,21 +120,11 @@ class TagsController extends Controller
    
             $tag = Tag::find($tag_id);
   
-  
-            if($tag) {
-              $tag->user()->detach();
-              $tag->posts()->detach();
-              $tag->projects()->detach();
-              $tag->tags()->detach();
-              $tag->delete();
-            }
+             if($tag) $tag->delete();
+            
 
           }
-          
-     
- 
-    
- 
+
          session()->flash('message', new MessageBag(['status' => 'success',
                                                      'message' => 'Yeah! All selected tags were deleted.']));
